move all tests to same package as base project
