+++ src/test/kotlin/com/github/thedeathlycow/calyx/kt/CycleTest.kt
@@ -1,4 +1,4 @@
-package com.github.calyx.kt.test
+package com.github.calyx.kt
 
 import com.github.thedeathlycow.calyx.kt.Cycle
 import com.github.thedeathlycow.calyx.kt.Options
