+++ src/test/kotlin/com/github/thedeathlycow/calyx/kt/FilterTest.kt
@@ -1,4 +1,4 @@
-package com.github.calyx.kt.test
+package com.github.calyx.kt
 
 import com.github.thedeathlycow.calyx.kt.*
 import org.junit.jupiter.api.Test
